restore from_center functiionality
This restores a feature of the layout-ios multiproduct that appears to have been lost when migrating to open-source. I've overloaded lyt_before and lyt_after to support a fromCenter parameter. If set to YES, the code will then test for whether a view is before or after the center of another view, allowing some overlapping. If NO, (equivalent to the original method), the view must strictly be before or after the edge of the other view - no overlapping.